+++ packages/zarm-cli/src/templates/component/compTpl.ts
@@ -2,21 +2,20 @@ import { paramCase } from 'change-case';
 
 export default (compName) => `import * as React from 'react';
 import classnames from 'classnames';
+import { ConfigContext } from '../n-config-provider';
+import type { Base${compName}Props } from './interface';
 
-export interface ${compName}Props {
-  prefixCls?: string;
+export interface ${compName}Props extends Base${compName}Props {
   className?: string;
 }
 
 const ${compName} = React.forwardRef<unknown, ${compName}Props>((props, ref) => {
-  const {
-    prefixCls,
-    className,
-    children,
-    ...restProps
-  } = props;
+  const { className, children, ...restProps } = props;
 
-  const compRef = (ref as any) || React.createRef<HTMLElement>();
+  const compRef = (ref as any) || React.createRef<HTMLDivElement>();
+
+  const { prefixCls: globalPrefixCls } = React.useContext(ConfigContext);
+  const prefixCls = \`\${globalPrefixCls}-${paramCase(compName)}\`;
   const cls = classnames(prefixCls, className);
 
   return (
@@ -28,9 +27,7 @@ const ${compName} = React.forwardRef<unknown, ${compName}Props>((props, ref) => 
 
 ${compName}.displayName = '${compName}';
 
-${compName}.defaultProps = {
-  prefixCls: 'za-${paramCase(compName)}',
-};
+${compName}.defaultProps = {};
 
 export default ${compName};