+++ packages/zarm/src/collapse/CollapseItem.tsx
@@ -1,105 +1,64 @@
-import React, { HTMLAttributes, PureComponent } from 'react';
+import React, { HTMLAttributes } from 'react';
 import classnames from 'classnames';
-import { BaseCollapseItemProps } from './PropsType';
+import type { BaseCollapseItemProps } from './interface';
+import { ConfigContext } from '../n-config-provider';
+import { useSafeLayoutEffect } from '../utils/hooks';
 
-export interface CollapseItemProps
-  extends Omit<HTMLAttributes<HTMLDivElement>, 'key' | 'title' | 'onChange'>,
-    BaseCollapseItemProps {
-  prefixCls?: string;
-}
+export type CollapseItemProps = Omit<HTMLAttributes<HTMLDivElement>, 'key' | 'title' | 'onChange'> &
+  BaseCollapseItemProps;
 
-interface CollapseItemStates {
-  active?: boolean;
-  prevActive?: boolean;
-}
+const CollapseItem = React.forwardRef<unknown, CollapseItemProps>((props, ref) => {
+  const { title, className, disabled, animated, isActive, children, onChange, ...rest } = props;
 
-export default class CollapseItem extends PureComponent<CollapseItemProps, CollapseItemStates> {
-  static defaultProps: CollapseItemProps = {
-    prefixCls: 'za-collapse-item',
-    animated: false,
-    disabled: false,
-  };
-
-  private content: HTMLDivElement | null;
-
-  state: CollapseItemStates = {
-    active: this.props.isActive,
-  };
-
-  static getDerivedStateFromProps(nextProps, state) {
-    if ('isActive' in nextProps && nextProps.isActive !== state.prevActive) {
-      return {
-        active: nextProps.isActive,
-        prevActive: nextProps.isActive,
-      };
-    }
-
-    return null;
-  }
-
-  componentDidMount() {
-    this.setStyle();
-  }
-
-  componentDidUpdate() {
-    this.setStyle();
-  }
-
-  onClickItem = () => {
-    const { onChange, disabled } = this.props;
-    const { active } = this.state;
+  const content = (ref as any) || React.createRef<HTMLElement>();
+  const collapseItemRef = (ref as any) || React.createRef<HTMLElement>();
+  const { prefixCls: globalPrefixCls } = React.useContext(ConfigContext);
+  const prefixCls = `${globalPrefixCls}-collapse-item`;
+  const onClickItem = () => {
     if (disabled) return;
-    typeof onChange === 'function' && onChange(active);
+    typeof onChange === 'function' && onChange(isActive!);
   };
 
-  setStyle = () => {
-    if (!this.content) return;
-
-    const { active } = this.state;
-    this.content.style.height = active ? `${this.getContentHeight(this.content)}px` : '0px';
-  };
-
-  getContentHeight = (content) => {
-    const children = [...content.children];
-    return children.reduce((res, next) => {
+  const getContentHeight = (ele) => {
+    const contentChildren = [...ele.children];
+    return contentChildren.reduce((res, next) => {
       res += next.offsetHeight;
       return res;
     }, 0);
   };
 
-  render() {
-    const {
-      prefixCls,
-      title,
-      className,
-      disabled,
-      animated,
-      isActive,
-      children,
-      onChange,
-      ...rest
-    } = this.props;
-    const { active } = this.state;
+  const setStyle = React.useCallback(() => {
+    if (!content.current) return;
+    content.current.style.height = isActive ? `${getContentHeight(content.current)}px` : '0px';
+  }, [content, isActive]);
 
-    const cls = classnames(prefixCls, className, {
-      [`${prefixCls}--active`]: active,
-      [`${prefixCls}--disabled`]: disabled,
-    });
-    return (
-      <div className={cls} {...rest}>
-        <div className={`${prefixCls}__header`} onClick={this.onClickItem}>
-          <div className={`${prefixCls}__title`}>{title}</div>
-          <div className={`${prefixCls}__arrow`} />
-        </div>
-        <div
-          className={`${prefixCls}__content`}
-          ref={(content) => {
-            this.content = content;
-          }}
-        >
-          <div className={`${prefixCls}__content__inner`}>{children}</div>
-        </div>
+  const cls = classnames(prefixCls, className, {
+    [`${prefixCls}--active`]: isActive,
+    [`${prefixCls}--disabled`]: disabled,
+  });
+
+  useSafeLayoutEffect(() => {
+    setStyle();
+  }, [setStyle]);
+
+  return (
+    <div className={cls} {...rest} ref={collapseItemRef}>
+      <div className={`${prefixCls}__header`} onClick={onClickItem}>
+        <div className={`${prefixCls}__title`}>{title}</div>
+        <div className={`${prefixCls}__arrow`} />
       </div>
-    );
-  }
-}
+      <div className={`${prefixCls}__content`} ref={content}>
+        <div className={`${prefixCls}__content__inner`}>{children}</div>
+      </div>
+    </div>
+  );
+});
+
+CollapseItem.displayName = 'CollapseItem';
+
+CollapseItem.defaultProps = {
+  animated: false,
+  disabled: false,
+};
+
+export default CollapseItem;