+++ packages/zarm/src/keyboard/Keyboard.tsx
@@ -17,10 +17,11 @@ const KEYS: { [type in KeyType]: readonly string[] } = {
 
 export interface KeyboardProps extends BaseKeyBoardProps {
   className?: string;
+  style?: React.CSSProperties;
 }
 
 const Keyboard = React.forwardRef<unknown, KeyboardProps>((props, ref) => {
-  const { className, type, onKeyClick } = props;
+  const { className, type, onKeyClick, ...restProps } = props;
   const keyboardRef = (ref as any) || React.createRef<HTMLDivElement>();
 
   const { locale: globalLocal, prefixCls: globalPrefixCls } = React.useContext(ConfigContext);
@@ -29,6 +30,9 @@ const Keyboard = React.forwardRef<unknown, KeyboardProps>((props, ref) => {
 
   const cls = classnames(prefixCls, className);
   const getKeys = KEYS[type!];
+
+  let touchLongPressTimer = 0;
+  let mouseLongPressTimer = 0;
 
   const onKeyPress = (e, key: string) => {
     e.stopPropagation();
@@ -40,6 +44,34 @@ const Keyboard = React.forwardRef<unknown, KeyboardProps>((props, ref) => {
     if (typeof onKeyClick === 'function') {
       onKeyClick(key);
     }
+  };
+
+  const onTouchLongPressIn = (e: React.TouchEvent<HTMLDivElement>, key: string) => {
+    onKeyPress(e, key);
+    touchLongPressTimer = window.setTimeout(() => {
+      touchLongPressTimer = window.setInterval(() => {
+        onKeyPress(e, key);
+      }, 100);
+    }, 800);
+  };
+
+  const onTouchLongPressOut = (e: React.TouchEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    window.clearInterval(touchLongPressTimer);
+  };
+
+  const onMouseLongPressIn = (e: React.MouseEvent<HTMLDivElement>, key: string) => {
+    onKeyPress(e, key);
+    mouseLongPressTimer = window.setTimeout(() => {
+      mouseLongPressTimer = window.setInterval(() => {
+        onKeyPress(e, key);
+      }, 100);
+    }, 800);
+  };
+
+  const onMouseLongPressOut = (e: React.MouseEvent<HTMLDivElement>) => {
+    e.preventDefault();
+    window.clearInterval(mouseLongPressTimer);
   };
 
   // todo: 长按连续删除还有点问题
@@ -66,12 +98,16 @@ const Keyboard = React.forwardRef<unknown, KeyboardProps>((props, ref) => {
   };
 
   return (
-    <div className={cls} ref={keyboardRef}>
+    <div className={cls} ref={keyboardRef} {...restProps}>
       <div className={`${prefixCls}__keys`}>{getKeys.map(renderKey)}</div>
       <div className={`${prefixCls}__handle`}>
         <div
           className={`${prefixCls}__item`}
-          onClick={(e) => onKeyPress(e, 'delete')}
+          onTouchStart={(e) => onTouchLongPressIn(e, 'delete')}
+          onTouchEnd={onTouchLongPressOut}
+          onTouchCancel={onTouchLongPressOut}
+          onMouseDown={(e) => onMouseLongPressIn(e, 'delete')}
+          onMouseUp={onMouseLongPressOut}
           // {...longPressEvent}
         >
           <DeleteKeyIcon size="lg" />