+++ packages/zarm/src/loading/interface.ts
@@ -1,0 +1,3 @@
+import type BaseLoadingProps from '../toast/interface';
+
+export type { BaseLoadingProps };