+++ packages/zarm/src/message/__tests__/index.test.tsx
@@ -4,7 +4,7 @@ import Message from '../index';
 import ConfigProvider from '../../n-config-provider';
 
 jest.mock('@zarm-design/icons', () => ({
-  ...jest.requireActual('@zarm-design/icons'),
+  ...(jest.requireActual('@zarm-design/icons') as any),
   Close: ({ onClick }) => (
     <div onClick={onClick} data-testid="close-icon">
       close icon