+++ packages/zarm/src/search-bar/locale/en_US.tsx
@@ -1,4 +1,0 @@
-export default {
-  placeholder: 'Search',
-  cancelText: 'Cancel',
-};